Made some edits.
This introduces two bugs, first is with the Inhabitant's eye symbol, the second is with the Nomaian plant pot. Both objects are lit strangely until highlighted or placed in darkness. The Nomaian plant pot is also textured strangely, and doesn't really look great. I'm not entirely sure what went wrong with it :(